Handle history checkout

+++ EcoPark/Components/PopUp/PartnerCheckOutPopUp/PartnerCheckOutPopUp.swift
@@ -26,7 +26,7 @@ class PartnerCheckOutPopUp: BasePopUpView {
         viewPopUp.btnCheckAgain.addTarget(self, action: #selector(btnCheckAgainTapped), for: .touchUpInside)
     }
     
-    func showPopUp(width: CGFloat, height: CGFloat, price: Int, vehicleType: String, vehicleNumber: String, checkOutNumber: String, completionCancel: CompletionClosure?, completionCheckAgain: CompletionClosure?, completionCheckout: CompletionClosure?) {
+    func showPopUp(width: CGFloat, height: CGFloat, price: Double, vehicleType: String, vehicleNumber: String, checkOutNumber: String, completionCancel: CompletionClosure?, completionCheckAgain: CompletionClosure?, completionCheckout: CompletionClosure?) {
         viewPopUp.lbPrice.text = price.description.formatNumber(type: ",")
         viewPopUp.lbVehicleType.text = vehicleType
         viewPopUp.lbVehicleNumber.text = vehicleNumber
